Fix compiler warnings
* Used "./configure --enable-warnings --enable-werror".
* Fixed all the warnings that appeared during 'make' in OS X 10.7
  using Apple GCC 4.2.
* Turned on more warnings in Xcode project.
* Fixed additional warnings turned up  by Clang 3.1.
* Enabled 32-bit builds in Xcode.
* Fixed some 64-to-32-bit implicit conversion warnings.

Reviewed-on: http://review.couchbase.org/15116
Tested-by: Damien Katz <damien@couchbase.com>

+++ include/libcouchstore/couch_db.h
@@ -152,7 +152,7 @@ extern "C" {
     couchstore_error_t couchstore_save_documents(Db *db,
                                                  Doc* const *docs,
                                                  DocInfo* const *infos,
-                                                 long numDocs,
+                                                 unsigned numDocs,
                                                  couchstore_save_options options);
     /**
      * Commit all pending changes and flush buffers to persistent storage.
