+++ tests/testapp.c
@@ -65,8 +65,8 @@ static void test_bitfield_fns(void)
     test_get_48(0xBA9876543210LL, expected4);
 }
 
-static void setdoc(Doc *doc, DocInfo *info, char *id, int idlen,
-                   char *data, int datalen, char *meta, int metalen)
+static void setdoc(Doc *doc, DocInfo *info, char *id, size_t idlen,
+                   char *data, size_t datalen, char *meta, size_t metalen)
 {
     memset(doc, 0, sizeof(Doc));
     memset(info, 0, sizeof(DocInfo));
@@ -467,7 +467,7 @@ static void mb5086(void)
 int main(int argc, const char *argv[])
 {
     int doc_counts[] = { 4, 69, 666, 9090, 99999, 666666 };
-    int i;
+    unsigned i;
     const char *small_doc_tpl = "{\"test_doc_index\":%d}";
     const char *large_doc_tpl =
         "{"
